AnswerService에 verifyAuthorizedMember() 추가

+++ backend/preproject/src/main/java/com/seb43_pre_12/preproject/answers/service/AnswerService.java
@@ -11,6 +11,9 @@ import com.seb43_pre_12.preproject.question.service.QuestionService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -41,6 +44,7 @@ public class AnswerService {
         return repository.save(answer);
     }
     public Answer updateAnswer(Answer answer) {
+        verifyAuthorizedMember(answer.getAnswerId());
         Answer findAnswer = findVerifedAnswer(answer.getAnswerId());
 
         Optional.ofNullable(answer.getContent()).ifPresent(content -> findAnswer.setContent(content));
@@ -64,6 +68,7 @@ public class AnswerService {
 
 
     public void deleteAnswer(long answerId) {
+        verifyAuthorizedMember(answerId);
         repository.delete(findVerifedAnswer(answerId));
     }
 
@@ -81,5 +86,18 @@ public class AnswerService {
 
     private Member verifyExistingMember(Member member) {
         return memberService.findVerifiedMember(member.getMemberId());
+    }
+    private void verifyAuthorizedMember(Long answerId) {
+        SecurityContext securityContext = SecurityContextHolder.getContext();
+        Authentication authentication = securityContext.getAuthentication();
+        String memberEmail = (String)authentication.getPrincipal();
+
+        // 댭변을 작성한 회원 객체를 찾는 로직
+        Answer verifedAnswer = findVerifedAnswer(answerId);
+        Member ownerOfAnswer = verifedAnswer.getMember();
+        final String AnswerOwnerEmail = ownerOfAnswer.getEmail();
+        // 댭변을 작성한 회원 객체의 email 과 로그인한 회원의 email 이 동일한지 조건문을 통해서 검사한다.
+        if(!memberEmail.equals(AnswerOwnerEmail)) throw  new BusinessLogicException(ExceptionCode.MEMBER_NOT_VALID);
+
     }
 }
